feat: add verifier, optimize quicknet verifier logic
