+++ pallets/drand/src/bls12_381.rs
@@ -1,77 +1,43 @@
-use crate::utils::ScalarFieldFor;
-use ark_bls12_381::{G1Affine, G1Projective, G2Affine, G2Projective};
-use ark_ec::{pairing::Pairing, short_weierstrass::SWCurveConfig, AffineRepr};
+/*
+ * Copyright 2024 by Ideal Labs, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use ark_ec::pairing::Pairing;
+use ark_std::{ops::Neg, Zero};
 use sp_ark_bls12_381::{
-	Bls12_381 as Bls12_381Opt, G1Affine as G1AffineOpt, G1Projective as G1ProjectiveOpt,
-	G2Affine as G2AffineOpt, G2Projective as G2ProjectiveOpt,
+	Bls12_381 as Bls12_381Opt, G1Affine as G1AffineOpt, G2Affine as G2AffineOpt,
 };
 
+/// An optimized way to verify Drand pulses from quicket
+/// Instead of computing two pairings and comparing them, we instead compute a multi miller loop,
+/// and then take the final exponentiation, saving a lot of computational cost.
+///
+/// This function is also inlined as a way to optimize performance.
+///
+/// * `signature`:
+/// * `q`:
+/// * `msg_on_curve`: The message signed by Drand, hashed to G1
+/// * `p_pub`: The beacon public key
 #[inline]
-pub fn pairing(a: G1Affine, b: G2Affine) {
-	let _out = ark_bls12_381::Bls12_381::multi_pairing([a], [b]);
-}
-
-#[inline]
-pub fn pairing_opt(a: G1AffineOpt, b: G2AffineOpt) {
-	let _out = Bls12_381Opt::multi_pairing([a], [b]);
-}
-
-#[inline]
-pub fn msm_g1(bases: &[G1Affine], scalars: &[ScalarFieldFor<G1Affine>]) {
-	let _out = <ark_bls12_381::g1::Config as SWCurveConfig>::msm(bases, scalars);
-}
-
-#[inline]
-pub fn msm_g1_opt(bases: &[G1AffineOpt], scalars: &[<G1AffineOpt as AffineRepr>::ScalarField]) {
-	let _out = <sp_ark_bls12_381::g1::Config as SWCurveConfig>::msm(bases, scalars);
-}
-
-#[inline]
-pub fn msm_g2(bases: &[G2Affine], scalars: &[ScalarFieldFor<G2Affine>]) {
-	let _out = <ark_bls12_381::g2::Config as SWCurveConfig>::msm(bases, scalars);
-}
-
-#[inline]
-pub fn msm_g2_opt(bases: &[G2AffineOpt], scalars: &[ScalarFieldFor<G2AffineOpt>]) {
-	let _out = <sp_ark_bls12_381::g2::Config as SWCurveConfig>::msm(bases, scalars);
-}
-
-#[inline]
-pub fn mul_projective_g1(base: &G1Projective, scalar: &[u64]) {
-	let _out = <ark_bls12_381::g1::Config as SWCurveConfig>::mul_projective(base, scalar);
-}
-
-#[inline]
-pub fn mul_projective_g1_opt(base: &G1ProjectiveOpt, scalar: &[u64]) {
-	let _out = <sp_ark_bls12_381::g1::Config as SWCurveConfig>::mul_projective(base, scalar);
-}
-
-#[inline]
-pub fn mul_affine_g1(base: &G1Affine, scalar: &[u64]) {
-	let _out = <ark_bls12_381::g1::Config as SWCurveConfig>::mul_affine(base, scalar);
-}
-
-#[inline]
-pub fn mul_affine_g1_opt(base: &G1AffineOpt, scalar: &[u64]) {
-	let _out = <sp_ark_bls12_381::g1::Config as SWCurveConfig>::mul_affine(base, scalar);
-}
-
-#[inline]
-pub fn mul_projective_g2(base: &G2Projective, scalar: &[u64]) {
-	let _out = <ark_bls12_381::g2::Config as SWCurveConfig>::mul_projective(base, scalar);
-}
-
-#[inline]
-pub fn mul_projective_g2_opt(base: &G2ProjectiveOpt, scalar: &[u64]) {
-	let _out = <sp_ark_bls12_381::g2::Config as SWCurveConfig>::mul_projective(base, scalar);
-}
-
-#[inline]
-pub fn mul_affine_g2(base: &G2Affine, scalar: &[u64]) {
-	let _out = <ark_bls12_381::g2::Config as SWCurveConfig>::mul_affine(base, scalar);
-}
-
-#[inline]
-pub fn mul_affine_g2_opt(base: &G2AffineOpt, scalar: &[u64]) {
-	let _out = <sp_ark_bls12_381::g2::Config as SWCurveConfig>::mul_affine(base, scalar);
+pub fn fast_pairing_opt(
+	signature: G1AffineOpt,
+	q: G2AffineOpt,
+	r: G1AffineOpt,
+	s: G2AffineOpt,
+) -> bool {
+	let looped = Bls12_381Opt::multi_miller_loop([signature.neg(), r], [q, s]);
+	let exp = Bls12_381Opt::final_exponentiation(looped);
+	exp.unwrap().is_zero()
 }
